test_sradb.py changes
